Organizes board page components into specific folders
